requirements.txt
newspaper3k
tweepy
pandas
nltk
spacy
matplotlib
wordcloud
